requirements.txt
# =============================================================================
# requirements.txt — Dependências do sistema de multiagentes
# =============================================================================
# Instale com: pip install -r requirements.txt

# ---------------------------------------------------------------------------
# OpenAI SDK (client oficial — GPT-4o)
# ---------------------------------------------------------------------------
openai>=1.51.0

# ---------------------------------------------------------------------------
# Variáveis de ambiente (.env)
# ---------------------------------------------------------------------------
python-dotenv>=1.0.0

# ---------------------------------------------------------------------------
# Requisições HTTP (Notion REST API)
# ---------------------------------------------------------------------------
requests>=2.31.0

# ---------------------------------------------------------------------------
# Agendamento do loop do Focus Guard
# ---------------------------------------------------------------------------
schedule>=1.2.1

# ---------------------------------------------------------------------------
# Utilitários de data/hora (já incluído no stdlib, mas explicitar versão)
# ---------------------------------------------------------------------------
# python-dateutil>=2.8.2   # descomente se precisar de parsing avançado de datas

# ---------------------------------------------------------------------------
# Colorama (compatibilidade de cores ANSI no Windows)
# ---------------------------------------------------------------------------
colorama>=0.4.6

# ---------------------------------------------------------------------------
# Opcional: rich (terminal mais bonito) — descomente para usar
# ---------------------------------------------------------------------------
# rich>=13.7.0
